requirements.txt
python-dotenv
streamlit
unstructured
langchain-chroma
langchain-groq
langchain-community
langchain-huggingface